requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.5
idna==2.10
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
